# backend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "malapos",
  "version": "0.0.1"
 },
 "paths": {
...
  "/api/v1/modifiers/product/{productId}": {
   "get": {
    "operationId": "getModifiersProductProductId",
    "summary": "Get a product",
    "tags": [
     "modifiers"
    ],
    "parameters": [
     {
      "name": "productId",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/modifiers.ts:208",
     "guards": [
      "requireAuth"
     ]
    }
   },
   "put": {
    "operationId": "putModifiersProductProductId",
    "summary": "Replace a product",
    "tags": [
     "modifiers"
    ],
    "parameters": [
     {
      "name": "productId",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/modifiers.ts:226",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "attachInput"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "groupIds": {
          "type": "array",
          "items": {
           "type": "string"
          }
         }
        },
        "required": [
         "groupIds"
        ],
        "additionalProperties": false
       }
      }
     }
    }
   }
  },
...
  "/api/v1/products": {
   "get": {
    "operationId": "getProducts",
    "summary": "List products",
    "tags": [
     "products"
    ],
    "parameters": [
     {
      "name": "active",
      "in": "query",
      "required": false,
      "schema": {}
     },
     {
      "name": "categoryId",
      "in": "query",
      "required": false,
      "schema": {}
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {}
     }
    ],
    "x-forjio": {
     "source": "src/routes/products.ts:30",
     "guards": [
      "requireAuth"
     ]
    }
   },
   "post": {
    "operationId": "postProducts",
    "summary": "Create a product",
    "tags": [
     "products"
    ],
    "parameters": [],
    "x-forjio": {
     "source": "src/routes/products.ts:81",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "productCreate"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "name": {
          "type": "string",
          "minLength": 1,
          "maxLength": 160
         },
         "description": {
          "type": "string",
          "maxLength": 2000,
          "nullable": true
         },
         "categoryId": {
          "type": "string",
          "nullable": true
         },
         "kind": {
          "type": "string",
          "enum": [
           "GOODS",
           "SERVICE"
          ],
          "default": "GOODS"
         },
         "trackStock": {
          "type": "boolean"
         },
         "requiresBatch": {
          "type": "boolean",
          "default": false
         },
         "imageUrl": {
          "type": "string",
          "maxLength": 600,
          "nullable": true
         },
         "isActive": {
          "type": "boolean",
          "default": true
         },
         "variants": {
          "type": "array",
          "items": {
           "type": "object",
           "properties": {
            "name": {
             "type": "string",
             "minLength": 1,
             "maxLength": 80,
             "default": "Default"
            },
            "sku": {
             "type": "string",
             "maxLength": 64,
             "nullable": true
            },
            "barcode": {
             "type": "string",
             "maxLength": 64,
             "nullable": true
            },
            "price": {
             "type": "integer",
             "minimum": 0,
             "maximum": 1000000000
            },
            "cost": {
             "type": "integer",
             "minimum": 0,
             "maximum": 1000000000,
             "default": 0
            },
            "sortOrder": {
             "type": "integer",
             "minimum": 0,
             "default": 0
            }
           },
           "required": [
            "price"
           ],
           "additionalProperties": false
          },
          "minItems": 1,
          "maxItems": 50
         }
        },
        "required": [
         "name",
         "variants"
        ],
        "additionalProperties": false
       }
      }
     }
    }
   }
  },
  "/api/v1/products/{id}": {
   "get": {
    "operationId": "getProductsId",
    "summary": "Get a product",
    "tags": [
     "products"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/products.ts:150",
     "guards": [
      "requireAuth"
     ]
    }
   },
   "patch": {
    "operationId": "patchProductsId",
    "summary": "Update a product",
    "tags": [
     "products"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/products.ts:163",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "productPatch"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "name": {
          "type": "string",
          "minLength": 1,
          "maxLength": 160
         },
         "description": {
          "type": "string",
          "maxLength": 2000,
          "nullable": true
         },
         "categoryId": {
          "type": "string",
          "nullable": true
         },
         "kind": {
          "type": "string",
          "enum": [
           "GOODS",
           "SERVICE"
          ]
         },
         "trackStock": {
          "type": "boolean"
         },
         "requiresBatch": {
          "type": "boolean"
         },
         "imageUrl": {
          "type": "string",
          "maxLength": 600,
          "nullable": true
         },
         "isActive": {
          "type": "boolean"
         },
         "variants": {
          "type": "array",
          "items": {
           "type": "object",
           "properties": {
            "id": {
             "type": "string",
             "minLength": 1
            },
            "name": {
             "type": "string",
             "minLength": 1,
             "maxLength": 80,
             "default": "Default"
            },
            "sku": {
             "type": "string",
             "maxLength": 64,
             "nullable": true
            },
            "barcode": {
             "type": "string",
             "maxLength": 64,
             "nullable": true
            },
            "price": {
             "type": "integer",
             "minimum": 0,
             "maximum": 1000000000
            },
            "cost": {
             "type": "integer",
             "minimum": 0,
             "maximum": 1000000000,
             "default": 0
            },
            "sortOrder": {
             "type": "integer",
             "minimum": 0,
             "default": 0
            }
           },
           "required": [
            "price"
           ],
           "additionalProperties": false
          },
          "minItems": 1,
          "maxItems": 50
         }
        },
        "additionalProperties": false
       }
      }
     }
    }
   },
   "delete": {
    "operationId": "deleteProductsId",
    "summary": "Delete a product",
    "tags": [
     "products"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/products.ts:257",
     "guards": [
      "requireAuth"
     ]
    }
   }
  },
  "/api/v1/products/{id}/variants": {
   "post": {
    "operationId": "postProductsIdVariants",
    "summary": "Variants a product",
    "tags": [
     "products"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/products.ts:284",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "variantInput"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "name": {
          "type": "string",
          "minLength": 1,
          "maxLength": 80,
          "default": "Default"
         },
         "sku": {
          "type": "string",
          "maxLength": 64,
          "nullable": true
         },
         "barcode": {
          "type": "string",
          "maxLength": 64,
          "nullable": true
         },
         "price": {
          "type": "integer",
          "minimum": 0,
          "maximum": 1000000000
         },
         "cost": {
          "type": "integer",
          "minimum": 0,
          "maximum": 1000000000,
          "default": 0
         },
         "sortOrder": {
          "type": "integer",
          "minimum": 0,
          "default": 0
         }
        },
        "required": [
         "price"
        ],
        "additionalProperties": false
       }
      }
     }
    }
   }
  },
  "/api/v1/products/{id}/variants/{vid}": {
   "patch": {
    "operationId": "patchProductsIdVariantsVid",
    "summary": "Update a variant",
    "tags": [
     "products"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "vid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/products.ts:310",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "variantInput.partial().extend({ isActive: z.boolean().optional() })"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "name": {
          "type": "string",
          "minLength": 1,
          "maxLength": 80,
          "default": "Default"
         },
         "sku": {
          "type": "string",
          "maxLength": 64,
          "nullable": true
         },
         "barcode": {
          "type": "string",
          "maxLength": 64,
          "nullable": true
         },
         "price": {
          "type": "integer",
          "minimum": 0,
          "maximum": 1000000000
         },
         "cost": {
          "type": "integer",
          "minimum": 0,
          "maximum": 1000000000,
          "default": 0
         },
         "sortOrder": {
          "type": "integer",
          "minimum": 0,
          "default": 0
         }
        },
        "additionalProperties": false
       }
      }
     }
    }
   },
   "delete": {
    "operationId": "deleteProductsIdVariantsVid",
    "summary": "Delete a variant",
    "tags": [
     "products"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "vid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/products.ts:335",
     "guards": [
      "requireAuth"
     ]
    }
   }
  },
  "/api/v1/products/{id}/variants/{vid}/recipe": {
   "get": {
    "operationId": "getProductsIdVariantsVidRecipe",
    "summary": "List recipe",
    "tags": [
     "products"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "vid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/products.ts:367",
     "guards": [
      "requireAuth"
     ]
    }
   },
   "put": {
    "operationId": "putProductsIdVariantsVidRecipe",
    "summary": "Replace-all the components of a variant + set its isComposite flag.",
    "description": "Replace-all the components of a variant + set its isComposite flag.",
    "tags": [
     "products"
    ],
    "parameters": [
     {
      "name": "id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     },
     {
      "name": "vid",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string"
      }
     }
    ],
    "x-forjio": {
     "source": "src/routes/products.ts:399",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "recipePut"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "isComposite": {
          "type": "boolean"
         },
         "components": {
          "type": "array",
          "items": {
           "type": "object",
           "properties": {
            "componentVariantId": {
             "type": "string",
             "minLength": 1
            },
            "quantity": {
             "type": "number",
             "exclusiveMinimum": true,
             "minimum": 0
            },
            "unit": {
             "type": "string",
             "maxLength": 40,
             "nullable": true
            }
           },
           "required": [
            "componentVariantId",
            "quantity"
           ],
           "additionalProperties": false
          },
          "maxItems": 100,
          "default": []
         }
        },
        "required": [
         "isComposite"
        ],
        "additionalProperties": false
       }
      }
     }
    }
   }
  },
  "/api/v1/products/bulk-category": {
   "post": {
    "operationId": "postProductsBulkCategory",
    "summary": "Create a bulk category",
    "tags": [
     "products"
    ],
    "parameters": [],
    "x-forjio": {
     "source": "src/routes/products.ts:132",
     "guards": [
      "requireAuth"
     ],
     "body": "validated",
     "bodySchema": "bulkCategorySchema"
    },
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "type": "object",
        "properties": {
         "productIds": {
          "type": "array",
          "items": {
           "type": "string",
           "minLength": 1
          },
          "minItems": 1,
          "maxItems": 500
         },
         "categoryId": {
          "type": "string",
          "minLength": 1,
          "nullable": true
         }
        },
        "required": [
         "productIds",
         "categoryId"
        ],
        "additionalProperties": false
       }
      }
     }
    }
   }
  },
  "/api/v1/products/lookup": {
   "get": {
    "operationId": "getProductsLookup",
    "summary": "Sell-screen lookup: exact barcode match first, else fuzzy name/sku.",
    "description": "Sell-screen lookup: exact barcode match first, else fuzzy name/sku.",
    "tags": [
     "products"
    ],
    "parameters": [
     {
      "name": "barcode",
      "in": "query",
      "required": false,
      "schema": {}
     },
     {
      "name": "q",
      "in": "query",
      "required": false,
      "schema": {}
     }
    ],
    "x-forjio": {
     "source": "src/routes/products.ts:50",
     "guards": [
      "requireAuth"
     ]
    }
   }
  },
...
 },
...
}